fix: convert modalCloseButton to Typescript
